refactor(model-postgres-json): Encapsulate query and sort compilation inside stateful PostgresJsonQueryCompiler
